requirements.txt
panda;
numpy;
scikit-learn;
xgboost;
seaborn;
matplotlib;
joblib;
streamlit;